[BE] refactor: library api 리팩토링
backend/model/ index & like album,tracks,playlists,artists 로 파일 분할
및
index.ts에서 병합
