+++ web/backend/models/library/index.ts
@@ -1,0 +1,6 @@
+import getUserLikeTracks from "./tracks";
+import getUserLikeAlbums from "./albums";
+import getUserLikeArtists from "./artists";
+import getUserLikePlaylists from "./playlists";
+
+export { getUserLikeAlbums, getUserLikeTracks, getUserLikeArtists, getUserLikePlaylists };
